........
  r17677 | bugman | 2012-10-02 16:35:26 +0200 (Tue, 02 Oct 2012) | 3 lines
  
  Python 3 integer division to float fix for the frame order analysis.
........
  r17678 | bugman | 2012-10-02 16:54:57 +0200 (Tue, 02 Oct 2012) | 5 lines
  
  Python 3 fix for the frame order system tests - float to string conversions behave differently.
  
  Now the explicit %.1f formatting is used to force only a single decimal place float.
........
  r17679 | bugman | 2012-10-02 16:58:40 +0200 (Tue, 02 Oct 2012) | 6 lines
  
  Python 3 fix for the frame order system tests matching r17678.
  
  As float to string conversions behave differently, the %.1f formatting is used to force only a
  single decimal place float.
........
  r17680 | bugman | 2012-10-02 16:59:22 +0200 (Tue, 02 Oct 2012) | 3 lines
  
  Python 3 fix - removed the use of the string.lower() function in the OpenDX mapping code.
........
  r17681 | bugman | 2012-10-02 17:47:23 +0200 (Tue, 02 Oct 2012) | 6 lines
  
  Python 3 fix for the writing out of XML formatted state and results files.
  
  In Python 3, the xmldoc.toprettyxml() returns a string object which needs to be 'encoded' prior to
  writing out to file.
........
  r17682 | bugman | 2012-10-02 17:58:04 +0200 (Tue, 02 Oct 2012) | 3 lines
  
  Python 3 import fixes for the generic_fns.structure package using relative paths.
........
  r17683 | bugman | 2012-10-02 18:29:25 +0200 (Tue, 02 Oct 2012) | 3 lines
  
  Python 3 fix for the format detection of results and save files.
........
  r17684 | bugman | 2012-10-02 18:34:02 +0200 (Tue, 02 Oct 2012) | 3 lines
  
  Python 2 fix for the relax_io.extract_data() function for a bug introduced at r17669.
........
  r17685 | bugman | 2012-10-02 18:46:16 +0200 (Tue, 02 Oct 2012) | 5 lines
  
  The relax_io.DummyFileObject now mimics a file object for both Python 2 and 3.
  
  In Python 3, everything from a file is of byte type and no longer string type.
........
  r17686 | bugman | 2012-10-02 18:48:25 +0200 (Tue, 02 Oct 2012) | 6 lines
  
  The Mf.test_write_results system test can now select the correct file to compare against in Python 3.
  
  The algorithm for determining if the 'final_results_trunc_1.3_v2' or
  'final_results_trunc_1.3_pre_py2.7.3_v2' file should be used could not handle Python 3.
........
  r17687 | bugman | 2012-10-02 18:55:31 +0200 (Tue, 02 Oct 2012) | 5 lines
  
  Another Python 3 fix for the relax_io.extract_data() function.
  
  Sometimes this function sees byte types, sometimes string types.  The function can now handle both.
........

+++ specific_fns/frame_order.py
@@ -1775,7 +1775,7 @@ class Frame_order(API_base, API_common):
             if cdp.params[i] in ['ave_pos_beta', 'eigen_beta', 'axis_theta']:
                 # Change the default increment numbers.
                 if not isinstance(inc, list):
-                    incs[i] = incs[i] / 2 + 1
+                    incs[i] = int(incs[i] / 2) + 1
 
                 # The distribution type and end point.
                 dist_type = 'acos'
